# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "podium API",
    "version": "0.1.0",
    "description": "REST API for the podium hackathon platform. Every UI action goes through these routes. Authorization is enforced server side and scoped to the event in the path. Request schemas are read from the same Zod validators the server runs.",
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "podium_session"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "operationId": "post_api_auth_register",
        "tags": [
          "auth"
        ],
        "summary": "POST /api/auth/register",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success. The body is JSON unless the route exports a file."
          },
          "400": {
            "description": "Validation failed. `error.details` maps field names to messages.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or hidden from this caller."
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 200
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  },
                  "accountType": {
                    "type": "string",
                    "enum": [
                      "user",
                      "organizer"
                    ]
                  }
                },
                "required": [
                  "email",
                  "password",
                  "name"
                ]
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "post_api_auth_login",
        "tags": [
          "auth"
        ],
        "summary": "POST /api/auth/login",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success. The body is JSON unless the route exports a file."
          },
          "400": {
            "description": "Validation failed. `error.details` maps field names to messages.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or hidden from this caller."
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        }
      }
    },
    "/api/auth/magic-link": {
      "post": {
        "operationId": "post_api_auth_magic_link",
        "tags": [
          "auth"
        ],
        "summary": "POST /api/auth/magic-link",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success. The body is JSON unless the route exports a file."
          },
          "400": {
            "description": "Validation failed. `error.details` maps field names to messages.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or hidden from this caller."
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  }
                },
                "required": [
                  "email"
                ]
              }
            }
          }
        }
      }
    },
    "/api/auth/magic-link/consume": {
      "post": {
        "operationId": "post_api_auth_magic_link_consume",
        "tags": [
          "auth"
        ],
        "summary": "POST /api/auth/magic-link/consume",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success. The body is JSON unless the route exports a file."
          },
          "400": {
            "description": "Validation failed. `error.details` maps field names to messages.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or hidden from this caller."
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 200
                  }
                },
                "required": [
                  "token"
                ]
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "post_api_auth_logout",
        "tags": [
          "auth"
        ],
        "summary": "POST /api/auth/logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success. The body is JSON unless the route exports a file."
          },
          "400": {
            "description": "Validation failed. `error.details` maps field names to messages.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or hidden from this caller."
          }
        }
      }
    },
    "/api/auth/session": {
      "get": {
        "operationId": "get_api_auth_session",
        "tags": [
          "auth"
        ],
        "summary": "GET /api/auth/session",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success. The body is JSON unless the route exports a file."
          },
          "400": {
            "description": "Validation failed. `error.details` maps field names to messages.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or hidden from this caller."
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "get_api_auth_me",
        "tags": [
          "auth"
        ],
        "summary": "GET /api/auth/me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success. The body is JSON unless the route exports a file."
          },
          "400": {
            "description": "Validation failed. `error.details` maps field names to messages.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in."
          },
          "403": {
            "description": "Signed in but not permitted."
          },
          "404": {
            "description": "Not found, or hidden from this caller."
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      },
      "patch": {
        "operationId": "patch_api_auth_me",
        "tags": [
          "auth"
        ],
        "summary": "PATCH /api/auth/me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success. The body is JSON unless the route exports a file."
          },
          "400": {
            "description": "Validation failed. `error.details` maps field names to messages.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in."
          },
          "403": {
            "description": "Signed in but not permitted."
          },
          "404": {
            "description": "Not found, or hidden from this caller."
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  },
                  "org": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 120
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "pronouns": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 40
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "bio": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 600
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "link": {
                    "anyOf": [
                      {
                        "type": "string",
                        "format": "uri",
                        "maxLength": 300
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "avatarHue": {
                    "type": "string",
                    "enum": [
                      "slate",
                      "blue",
                      "teal",
                      "green",
                      "amber",
                      "plum",
                      "rose",
                      "cyan",
                      "coral"
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/password": {
      "post": {
        "operationId": "post_api_auth_password",
        "tags": [
          "auth"
        ],
        "summary": "POST /api/auth/password",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success. The body is JSON unless the route exports a file."
          },
          "400": {
            "description": "Validation failed. `error.details` maps field names to messages.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in."
          },
          "403": {
            "description": "Signed in but not permitted."
          },
          "404": {
            "description": "Not found, or hidden from this caller."
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "currentPassword": {
                    "type": "string",
                    "minLength": 1
                  },
                  "newPassword": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 200
                  }
                },
                "required": [
                  "currentPassword",
                  "newPassword"
                ]
              }
            }
          }
        }
      }
    },
    "/api/auth/sessions/revoke": {
      "post": {
        "operationId": "post_api_auth_sessions_revoke",
        "tags": [
          "auth"
        ],
        "summary": "POST /api/auth/sessions/revoke",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success. The body is JSON unless the route exports a file."
          },
          "400": {
            "description": "Validation failed. `error.details` maps field names to messages.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in."
          },
          "403": {
            "description": "Signed in but not permitted."
          },
          "404": {
            "description": "Not found, or hidden from this caller."
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/api/auth/sessions": {
      "get": {
        "operationId": "get_api_auth_sessions",
        "tags": [
          "auth"
        ],
        "summary": "GET /api/auth/sessions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success. The body is JSON unless the route exports a file."
          },
          "400": {
            "description": "Validation failed. `error.details` maps field names to messages.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in."
          },
          "403": {
            "description": "Signed in but not permitted."
          },
          "404": {
            "description": "Not found, or hidden from this caller."
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/api/auth/sessions/{sessionId}": {
      "delete": {
        "operationId": "delete_api_auth_sessions_sessionId",
        "tags": [
          "auth"
        ],
        "summary": "DELETE /api/auth/sessions/{sessionId}",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success. The body is JSON unless the route exports a file."
          },
          "400": {
            "description": "Validation failed. `error.details` maps field names to messages.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in."
          },
          "403": {
            "description": "Signed in but not permitted."
          },
          "404": {
            "description": "Not found, or hidden from this caller."
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/api/auth/me/activity": {
      "get": {
        "operationId": "get_api_auth_me_activity",
        "tags": [
          "auth"
        ],
        "summary": "GET /api/auth/me/activity",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success. The body is JSON unless the route exports a file."
          },
          "400": {
            "description": "Validation failed. `error.details` maps field names to messages.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in."
          },
          "403": {
            "description": "Signed in but not permitted."
          },
          "404": {
            "description": "Not found, or hidden from this caller."
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/api/auth/me/notifications": {
      "get": {
        "operationId": "get_api_auth_me_notifications",
        "tags": [
          "auth"
        ],
        "summary": "GET /api/auth/me/notifications",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success. The body is JSON unless the route exports a file."
          },
          "400": {
            "description": "Validation failed. `error.details` maps field names to messages.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in."
          },
          "403": {
            "description": "Signed in but not permitted."
          },
          "404": {
            "description": "Not found, or hidden from this caller."
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      },
      "patch": {
        "operationId": "patch_api_auth_me_notifications",
        "tags": [
          "auth"
        ],
        "summary": "PATCH /api/auth/me/notifications",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success. The body is JSON unless the route exports a file."
          },
          "400": {
            "description": "Validation failed. `error.details` maps field names to messages.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in."
          },
          "403": {
            "description": "Signed in but not permitted."
          },
          "404": {
            "description": "Not found, or hidden from this caller."
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "judgingReminders": {
                    "type": "boolean"
                  },
                  "resultsPublished": {
                    "type": "boolean"
                  },
                  "voteDigest": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/tokens/{tokenId}": {
      "delete": {
        "operationId": "delete_api_auth_tokens_tokenId",
        "tags": [
          "auth"
        ],
        "summary": "DELETE /api/auth/tokens/{tokenId}",
        "parameters": [
          {
            "name": "tokenId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success. The body is JSON unless the route exports a file."
          },
          "400": {
            "description": "Validation failed. `error.details` maps field names to messages.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in."
          },
          "403": {
            "description": "Signed in but not permitted."
          },
          "404": {
            "description": "Not found, or hidden from this caller."
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
  }
}